Add player submission modified
Modified the submit() function when adding a player to the database and tree, specifically the tree. When the user adds data and the program repacks the main_frame, the tree table will now display the new data in the correct alphabetical position in the table, as well as highlight it to show the user where and what was stored.